fix: repo structure and runtime refactor
