+++ src/cli/base.rs
@@ -1,13 +1,8 @@
-use crate::{
-  core::{
-    core_services::nya_core::get_core_services, payload::Payload, service::Service
-  }, 
-  runtime::nya::Nya,
-  utils::{self, ConfigStatus}
-};
+use crate::core::runtime::Nya;
+use crate::utils::{resolve_base_config, ConfigStatus};
 
 pub async fn build(config: Option<String>) {
-  let valid_path = utils::resolve_base_config(config.as_deref());
+  let valid_path = resolve_base_config(config.as_deref());
   let path = match valid_path {
     ConfigStatus::Exists(path) => path,
     ConfigStatus::Missing(path) => {
@@ -15,13 +10,11 @@ pub async fn build(config: Option<String>) {
       return;
     }
   };
-  let services: Vec<Box<dyn Service>> = get_core_services();
-  let nya = Nya::build("base:build", vec![&path.display().to_string()], services);
-  nya.run(Payload::empty()).await;
+  Nya::run("base:build", vec![&path.display().to_string()]).await;
 }
 
 pub async fn destroy(config: Option<String>) {
-  let valid_path = utils::resolve_base_config(config.as_deref());
+  let valid_path = resolve_base_config(config.as_deref());
   let path = match valid_path {
     ConfigStatus::Exists(path) => path,
     ConfigStatus::Missing(path) => {
@@ -29,7 +22,5 @@ pub async fn destroy(config: Option<String>) {
       return;
     }
   };
-  let services: Vec<Box<dyn Service>> = get_core_services();
-  let nya = Nya::build("base:destroy", vec![&path.display().to_string()], services);
-  nya.run(Payload::empty()).await;
+  Nya::run("base:destroy", vec![&path.display().to_string()]).await;
 }
