+++ src/cli/ship.rs
@@ -1,6 +1,6 @@
 use colored::Colorize;
 
-use crate::{core::{core_services::nya_core::get_core_services, payload::Payload, service::Service}, runtime::nya::Nya, utils::{ConfigStatus, resolve_base_config, resolve_capsule}};
+use crate::{core::runtime::Nya, utils::{ConfigStatus, resolve_base_config, resolve_capsule}};
 
 pub async fn run(config: Option<String>, capsule: Option<String>) {
   let config_result = resolve_base_config(config.as_deref());
@@ -24,10 +24,8 @@ pub async fn run(config: Option<String>, capsule: Option<String>) {
   let nya_base_config_string = nya_base_config_path.display().to_string();
   let nya_capsule_string = nya_capsule_path.display().to_string();
   let context_file_path: Vec<&str> = vec![&nya_base_config_string, &nya_capsule_string];
-  let services: Vec<Box<dyn Service>> = get_core_services();
-  let nya = Nya::build("capsule:ship", context_file_path, services);
-  let _ = &nya.set("config_path", &nya_base_config_string).await;
-  let _ = &nya.set("capsule_path", &nya_capsule_string).await;
-  nya.run(Payload::empty()).await;
+  // TODO: these strings are still needed for the ship operation, 
+  // we need to either be able to set them or the operation needs to derive the values
+  Nya::run("capsule:ship", context_file_path).await;
 
 }
