+++ src/core/core_schema.json
