--- src/core/core_services.rs
@@ -1,3 +1,0 @@
-pub mod nya_core;
-pub mod nya_base;
-pub mod nya_ship;
