+++ src/lib.rs
@@ -1,6 +1,7 @@
 pub mod core;
-pub mod runtime;
 pub mod cli;
-pub mod embedded;
 pub mod utils;
 pub mod defaults;
+mod ops;
+mod external;
+
