+++ src/ops/base_build/templates/registries.yaml.j2
