+++ src/ops/pack/values.fullstack.yaml
