--- src/runtime.rs
@@ -1,1 +1,0 @@
-pub mod nya;
